Move the new incomplete command line parser to Core
