+++ UWP/CoreUWP/CoreUWP.vcxproj
@@ -86,6 +86,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="..\..\Core\AVIDump.h" />
+    <ClInclude Include="..\..\Core\CmdLine.h" />
     <ClInclude Include="..\..\Core\Compatibility.h" />
     <ClInclude Include="..\..\Core\Config.h" />
     <ClInclude Include="..\..\Core\ConfigSettings.h" />
@@ -359,6 +360,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\Core\AVIDump.cpp" />
+    <ClCompile Include="..\..\Core\CmdLine.cpp" />
     <ClCompile Include="..\..\Core\Compatibility.cpp" />
     <ClCompile Include="..\..\Core\Config.cpp" />
     <ClCompile Include="..\..\Core\ConfigSettings.cpp" />
